requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.29.0
pydantic==2.7.4
boto3==1.34.84        # S3 연동용 
pyhwp
pymupdf==1.25.5
six>=1.16.0
langchain==0.3.23
python-dotenv